Updated Architecture
