+++ .idea/.name
@@ -1,1 +1,1 @@
-Gaming Bets
+GamingBets
