+++ app/src/main/java/de/blogsiteloremipsum/gamingbets/client/ClientMethods.java
@@ -1,0 +1,24 @@
+package de.blogsiteloremipsum.gamingbets.client;
+
+import java.util.Date;
+
+import de.blogsiteloremipsum.gamingbets.classes.Bet;
+import de.blogsiteloremipsum.gamingbets.classes.User;
+
+/**
+ * Created by Felix Morsbach on 13.11.2015.
+ */
+public interface ClientMethods {
+
+    public boolean login (User user);
+
+    public boolean logout(User user);
+
+    public boolean register (String username, String email, String pw, Date dob);
+
+    public boolean edit(User user);
+
+    public boolean editAdmin(User user);
+
+    public boolean placeBet(Bet bet);
+}
