+++ app/src/main/java/de/blogsiteloremipsum/gamingbets/client/LocalClientSocket.java
@@ -1,41 +1,22 @@
-package de.blogsiteloremipsum.gamingbets.server;
+package de.blogsiteloremipsum.gamingbets.client;
 
-import java.util.ArrayList;
 import java.util.Date;
 
 import de.blogsiteloremipsum.gamingbets.classes.Bet;
 import de.blogsiteloremipsum.gamingbets.classes.User;
 
 /**
- * Created by Felix Morsbach on 13.11.2015.
+ * Created by Felix Morsbach on 16.11.2015.
  */
-public class Logic implements ServerMethods {
-
-    private ArrayList<User> users;
-
-    public Logic() {
-        users = new ArrayList<>();
-    }
-
+public class LocalClientSocket implements ClientMethods{
     @Override
     public boolean login(User user) {
-        //TODO overright isequal method in User!
-        if (!users.contains(user)) {
-            users.add(user);
-            return true;
-        } else {
-            return false;
-        }
+        return false;
     }
 
     @Override
     public boolean logout(User user) {
-        if (users.contains(user)) {
-            users.remove(user);
-            return true;
-        } else {
-            return false;
-        }
+        return false;
     }
 
     @Override
@@ -57,4 +38,7 @@ public class Logic implements ServerMethods {
     public boolean placeBet(Bet bet) {
         return false;
     }
+
+    //TODO Insert Socket!
+    //TODO Insert Communication Protocol
 }
